Separate into parse and plot
